# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Royal SL E-commerce API",
        "description": "REST API for the Royal SL Ecommerce System — admin panel, customer storefront, POS, warehouse & inventory management, coupons, payments, and reporting. All sales channels (website, Facebook, outlet POS) share a single central inventory.",
        "contact": {
            "email": "support@royalsl.example"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/admin/payment-methods": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "List available payment methods (cash, card, bank transfer, COD, online gateway)",
                "operationId": "c584e84e7498925c8356f7117fbcab59",
                "responses": {
                    "200": {
                        "description": "Payment method list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/roles": {
            "get": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "List roles with their permissions",
                "operationId": "1fbe54def12159b19db06ae306e57c85",
                "responses": {
                    "200": {
                        "description": "Role list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "List users (staff & customers)",
                "operationId": "80c2e09d986ac49741d14d9fe9e3f35b",
                "responses": {
                    "200": {
                        "description": "User list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Create a staff/admin user and assign roles",
                "operationId": "1e82bed5b21b4aa665386e9b5fd9905c",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/customer/addresses": {
            "get": {
                "tags": [
                    "Customer Orders"
                ],
                "summary": "List the customer's saved addresses",
                "operationId": "e76760fb73627a7108a1672f825adfd6",
                "responses": {
                    "200": {
                        "description": "Address list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Customer Orders"
                ],
                "summary": "Add a new address",
                "operationId": "e8294e9f7d48739b6fdf68b3c2b25271",
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/customer/orders": {
            "get": {
                "tags": [
                    "Customer Orders"
                ],
                "summary": "List the authenticated customer's orders",
                "operationId": "4d8843ad2a492e1c8f91efff05a64e7d",
                "responses": {
                    "200": {
                        "description": "Order list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/customer/checkout": {
            "post": {
                "tags": [
                    "Customer Orders"
                ],
                "summary": "Place a website order from the current cart / provided items",
                "operationId": "1dd3629b71e28be27876097bdef913ad",
                "responses": {
                    "201": {
                        "description": "Order placed"
                    },
                    "422": {
                        "description": "Insufficient stock or invalid coupon"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/inventory/stock": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "List current stock levels (optionally filtered by warehouse)",
                "operationId": "9448d9570ac3dfb71cc47df19399e71f",
                "parameters": [
                    {
                        "name": "warehouse_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stock list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/inventory/stock/low-stock": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "List active low-stock alerts",
                "operationId": "66471be345a20b0a7e36e70e7099d8fa",
                "responses": {
                    "200": {
                        "description": "Low stock alert list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/inventory/stock/movements": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "List the stock movement ledger (audit trail for every inventory change)",
                "operationId": "c6511802094dae7091927614958b3def",
                "responses": {
                    "200": {
                        "description": "Movement list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/inventory/suppliers": {
            "get": {
                "tags": [
                    "Inventory"
                ],
                "summary": "List suppliers",
                "operationId": "d35dec16c8bed5a18c9e92bf895a87c3",
                "responses": {
                    "200": {
                        "description": "Supplier list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Sanctum personal access token. Send as: Authorization: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}